gettimeofday has portability issues, see for example #11558.

Regardless of large-scale clock refactors in the future, I think it's fine for bench to just use std::chrono itself.

Note that this may slightly improve bench accuracy and changes the display from tiny floats to nanosecond counts instead.