@laanwj 2017-10-26
Does this also use a monotonic clock? That's another issue with using plain gettimeofday for benchmarks.