@laanwj 2017-10-26
> @laanwj Yes, if the resolution is no worse than the high precision clock. See the last commit: e2dc48c

Nice! concept ACK