@martinus 2017-10-27
I've also switched to std::chrono in https://github.com/bitcoin/bitcoin/pull/11517